refactore: SampleBuilderMacro+Struct and SampleBuilderMacro+Class
